IT WORKS. I DON'T KNOW WHAT I DID BUT IT WORKS.
Also, haven't been committing as I go, so this is all together in
one big blob. But screw it. It works. Somehow, all this illegible
spaghetti code works. The email stuff is held on with duct tape. I
have no idea why that's working. At some point I also apparently
gave up on modularity. It's a rough prototype, it doesn't matter.
What matters is that it works.

If a Kenco employee is reading this, I am so so sorry. It's
entirely possible that your boss is forcing you to use my code. If
that's the case, you probably think I'm a moron. I am, but that's
not the point. I was rushed, I took sloppy shortcuts. It works, but
as a prototype. A very very rough prototype. Really, I'm normally
better than this. Trust me. Anyways, this is going on too long,
I'm mildly sleep-deprived, I have to present this tomorrow, and
I still have more work to do.

Ryan, signing off.